pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "auren-forge"
version = "0.1.0"
description = "AI-orchestrated supervised fine-tuning for language models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Auren Research"}]
keywords = ["llm", "fine-tuning", "sft", "ai", "training"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "datasets>=2.14.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
train = [
    "unsloth",
    "trl>=0.7.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "bitsandbytes",
    "peft",
    "accelerate",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.10",
]

[project.scripts]
forge = "forge.cli:app"

[project.urls]
Homepage = "https://github.com/meryyllea/auren-forge"
Repository = "https://github.com/meryyllea/auren-forge"
Issues = "https://github.com/meryyllea/auren-forge/issues"

[tool.hatch.build.targets.wheel]
packages = ["forge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
